Split afa and common cypress tests
